requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.36
pydantic-settings==2.4.0
python-multipart==0.0.9
aiosqlite==0.20.0
loguru==0.7.2
python-dotenv==1.0.1